master

+++ app/Http/Livewire/App/DemandLikes.php
@@ -12,26 +12,31 @@ class DemandLikes extends Component
 
     public $like;
     public $demands;
+    public $like_id;
 
     public function mount(Demands $demands)
     {
         $this->demands = $demands;
+        $this->like_id = $demands->id;
     }
     public function render()
     {
         return view('livewire.app.demand-likes');
     }
-    public function like($id)
+    public function like()
     {
-        $this->like = DemandLike::where('demand_id',$id)->where('user_id',Auth::user()->id)->first();
+        if ($this->like_id ) {
+            $this->like = DemandLike::where('demand_id',$this->like_id)->where('user_id',Auth::user()->id)->first();
 
-        if($this->like != null){
-            $this->like->delete();
-        }else{
-            $this->like = DemandLike::create([
-                'demand_id' => $id,
-                'user_id'   => Auth::user()->id,
-            ]);
+            if($this->like != null){
+                $this->like->delete();
+            }else{
+                $this->like = DemandLike::create([
+                    'demand_id' => $this->like_id,
+                    'user_id'   => Auth::user()->id,
+                ]);
+            }
         }
+
     }
 }
